simple-rag/requirements.txt
qdrant-client==1.7.0
scikit-learn==1.3.0
numpy==1.24.3
langchain==0.1.0
langchain-community==0.0.10
pypdf==3.17.4
python-dotenv==1.0.0
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
tiktoken==0.5.2
opik==0.1.0
streamlit==1.29.0
openai==1.95.0